[qmf-notifications-plugin] Update to libaccounts-qt version 1.13
This commit ensures that we always use the Accounts::Account::fromId()
function instead of the Accounts::Manager::account() function to
create account instances, since we manually control the lifetime of
those instances.

The Accounts::Manager::account() behaviour was modified in
libaccounts-qt version 1.7 such that deleting the returned account
is unsafe.